new update

+++ src/main/java/eu/spitfire_project/smart_service_proxy/core/SelfDescription.java
@@ -89,7 +89,7 @@ public class SelfDescription {
 
             Long mediaTypeNumber = contentTypeOption.getDecodedValue();
             OptionRegistry.MediaType mediaType =
-                    OptionRegistry.MediaType.getByNumber(mediaTypeNumber.intValue());
+                    OptionRegistry.MediaType.getByNumber((long) mediaTypeNumber.intValue());
 
             if(mediaType == null){
                 log.info("[SelfDescription] CoapResponse contains an unknown content type (number = " +
